+++ cypress/tests/core/listing-template.js
@@ -73,4 +73,66 @@ describe('Folder Contents Tests', () => {
       .should('have.attr', 'aria-label')
       .and('eq', 'Play or Pause Slideshow');
   });
+
+  it('should only show pages if they are released and public', () => {
+    cy.visit('/my-folder/my-document');
+    cy.get('button.more').click();
+    cy.get('.state-select .react-select-container').click();
+    cy.get('div .react-select__option:contains("Public")').click();
+    cy.visit('/my-folder/my-document/edit');
+    cy.get('.block-editor-text').click();
+    cy.get('button.block-add-button').click();
+    cy.get('.mostUsed button.button.listing:first-of-type').click();
+    cy.get('#toolbar-save').click();
+    cy.get('.add').click();
+    cy.get('#toolbar-add-document').click();
+    cy.get('.block-editor-title').click().type('Page-1');
+    cy.get('#effective-date').click();
+    cy.get('.CalendarDay__today').click();
+    cy.get('#toolbar-save').click();
+    cy.get('button.more').click();
+    cy.get('.state-select .react-select-container').click();
+    cy.get('div .react-select__option:contains("Public")').click();
+    cy.get('.add').click();
+    cy.get('#toolbar-add-document').click();
+    cy.get('.block-editor-title').click().type('Page-2');
+    cy.get('#effective-date').click();
+    cy.get('.CalendarDay__today').click();
+    cy.get('#toolbar-save').click();
+    cy.visit('/my-folder/my-document');
+    cy.get('.add').click();
+    cy.get('#toolbar-add-document').click();
+    cy.get('.block-editor-title').click().type('Page-3');
+    cy.get('#effective-date').click();
+    cy.get('.CalendarDay__today').click();
+    cy.get('#effective-time').click();
+    cy.get(
+      '.rc-time-picker-panel-combobox > .rc-time-picker-panel-select:first-of-type .rc-time-picker-panel-select-option-selected +li',
+    ).click();
+    cy.get('#toolbar-save').click();
+    cy.get('button.more').click();
+    cy.get('.state-select .react-select-container').click();
+    cy.get('div .react-select__option:contains("Public")').click();
+    cy.get('.add').click();
+    cy.get('#toolbar-add-document').click();
+    cy.get('.block-editor-title').click().type('Page-4');
+    cy.get('#effective-date').click();
+    cy.get('.CalendarDay__today').click();
+    cy.get('#effective-time').click();
+    cy.get(
+      '.rc-time-picker-panel-combobox > .rc-time-picker-panel-select:first-of-type .rc-time-picker-panel-select-option-selected +li +li +li',
+    ).click();
+    cy.get('#toolbar-save').click();
+    cy.get('button.more').click();
+    cy.get('.state-select .react-select-container').click();
+    cy.get('div .react-select__option:contains("Public")').click();
+    cy.get('.user').click();
+    cy.get('#toolbar-logout').click();
+
+    cy.visit('/my-folder/my-document');
+    cy.get('h4').contains('Page-1').should('exist');
+    cy.get('h4').contains('Page-2').should('not.exist');
+    cy.get('h4').contains('Page-3').should('not.exist');
+    cy.get('h4').contains('Page-4').should('not.exist');
+  });
 });